refactor(web): split pages into server/client components and fix review findings
- Convert all 6 page.tsx files to server components (Principle XI)
- Extract client logic into *-form.tsx and *-content.tsx files
- Replace raw <a> tags with Link from @infrastructure/navigation
- Extract shared useSignOut hook to deduplicate sign-out logic
- Add error handling to sign-out flow with toast notifications
- Remove unsafe redirects from sign-up and reset-password forms
- Gate demo credentials behind NODE_ENV === "development"
- Move sonner to pnpm catalog with exact version (Principle V)
- Clean up commented-out code, use TODO comments consistently
- Update tests to match new component structure
