+++ apps/web/app/(auth)/sign-up/__tests__/page.test.tsx
@@ -4,24 +4,8 @@ import { render, screen } from "@testing-library/react";
 import type React from "react";
 import { describe, expect, it, vi } from "vitest";
 
-vi.mock("@infrastructure/supabase/auth", () => ({
-  useAuth: () => ({
-    signUp: vi.fn(),
-    signInWithOAuth: vi.fn(),
-    isLoading: false,
-    session: null,
-    user: null,
-    signIn: vi.fn(),
-    signOut: vi.fn(),
-  }),
-}));
-
 vi.mock("@infrastructure/navigation", () => ({
-  useNavigation: () => ({
-    navigate: vi.fn(),
-    replace: vi.fn(),
-    back: vi.fn(),
-  }),
+  Link: ({ children, ...props }: React.ComponentProps<"a">) => <a {...props}>{children}</a>,
 }));
 
 vi.mock("@/components/auth-form", () => ({
@@ -42,21 +26,21 @@ vi.mock("@/components/auth-form", () => ({
   ),
 }));
 
-import SignUpPage from "../page";
+import { SignUpForm } from "../sign-up-form";
 
-describe("SignUpPage", () => {
+describe("SignUpForm", () => {
   it("renders AuthForm with sign-up title", () => {
-    render(<SignUpPage />);
+    render(<SignUpForm />);
     expect(screen.getByText("Sign Up")).toBeDefined();
   });
 
   it("renders Create Account submit label", () => {
-    render(<SignUpPage />);
+    render(<SignUpForm />);
     expect(screen.getByText("Create Account")).toBeDefined();
   });
 
   it("renders sign-in link", () => {
-    render(<SignUpPage />);
+    render(<SignUpForm />);
     expect(screen.getByText(/Sign In/)).toBeDefined();
   });
 });
