+++ apps/web/package.json
@@ -25,7 +25,7 @@
     "next": "catalog:",
     "react": "catalog:",
     "react-dom": "catalog:",
-    "sonner": "^2.0.7"
+    "sonner": "catalog:"
   },
   "devDependencies": {
     "@infrastructure/typescript-config": "workspace:*",
