+++ pnpm-workspace.yaml
@@ -35,6 +35,7 @@ catalog:
 
   # UI Components
   "@base-ui/react": "1.2.0"
+  sonner: "2.0.7"
   "@radix-ui/react-slot": "1.1.1"
   lucide-react: "0.564.0"
